Big Update - Keys, Doors, Updated Enemy
Keys -
Keys can now be collected, Using the key script and box colliders, when the player is in the box collider it allows keys to be picked up. Key manager script tracks the number of keys a player has
Door -
Door script allows player to complete the game if they have collected all the scripts. It manipulates the alpha of a panel called GameComplete screen which turns the screen gradually more white and shows a text prompt on the UI at the bottom.
It uses same pickup mechanic as the keys and batteries (box collider)
If the player does not have enough keys it will tell them to collect keys using a text prompt on the UI. This is done by checking the key manager script to see how many keys have been collected.

Enemy -
Enemy now flickers randomly through manipulation of the renderer. Once invisible the enemy moves faster. This is done through using a couroutine. Also debugged enemy animations.