Added Improved Dialogue
When spotted or heard, the player will now say dialogue as text on the UI. Tried to improve collection text for the battery and medkit however bugs are still present. Adjusted the collection for items by adding a check for E pressed in the update method allowing for instant pickup of items instead of the old method which relied on the player triggering the collider of the object which caused delayed item pickup.