Added garages
Added garages where you can hide and interact with the doors. Made the animations using unity for the door opening and closing. Also wrote the door controller script which handles the different animation states and the doorinteraction script which when the player is in range allows them to toggle the door open and close. It is similar to the battery and medkit collection scripts