Integrated Dictation, player can now talk to NPCs using microphone
Using Windows.Speech I created a feature which would convert the players speech into text and display it on the screen. If within range of an NPC, and the player says certain keywords, the NPC will be able to reply back. Initially this caused problems with the microphone threshold feature which took in the player's loudness and if too loud alerted the monster to the player's location. I decided to implement a microphone manager which controlled all the microphone inputs, however this caused the microphone threshold feature to stop working, so I reverted to having it being directly called in the Enemy AI script.